+++ src/presentation/controllers/create-subtitle/create-subtitle-protocols.ts
@@ -1,0 +1,5 @@
+export { badRequest, notFound } from '../../helpers/http-helper'
+export { Controller, HttpRequest, HttpResponse } from '../../protocols'
+export { FileModel } from '../../../domain/models/file'
+export { GetFile } from '../../../domain/usecases/get-file'
+export { CreateSubtitle, CreateSubtitleModel } from '../../../domain/usecases/create-subtitle'
