+++ src/presentation/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './invalid-param-error'
 export * from './missing-param-error'
 export * from './server-error'
+export * from './not-found-error'
